Move keyboard functionality to custom hook
