+++ src/components/molecules/Combobox/Combobox.tsx
@@ -1,8 +1,9 @@
-import { useEffect, useId, useRef, useState } from "react";
+import { useId, useState } from "react";
 import { Listbox } from "./Listbox";
 import { Textbox } from "./Textbox";
 import type { ComboboxProps } from "./type";
 import { RequiredBadge } from "@components/atoms/RequiredBadge";
+import { useKeyboard } from "./hooks";
 
 export const Combobox = ({
   options,
@@ -16,96 +17,14 @@ export const Combobox = ({
   const [showList, setShowList] = useState(false);
   const [value, setValue] = useState("");
   const [selectedElementId, setSelectedElementId] = useState("");
-  const [isFocus, setIsFocus] = useState(false);
-  const [visualFocusIndex, setVisualFocusIndex] = useState<number | null>(null);
   const inputLabelId = useId();
-
-  useEffect(() => {
-    if (!isFocus) {
-      return;
-    }
-
-    const handleClosedListKey = (event: KeyboardEvent) => {
-      const key = event.key;
-
-      switch (key) {
-        case "Up":
-        case "ArrowUp":
-          event.preventDefault();
-          const lastIndex = options.length - 1;
-          setShowList(true);
-          setVisualFocusIndex(lastIndex);
-          return;
-        case "Down":
-        case "ArrowDown":
-          event.preventDefault();
-          const firstIndex = 0;
-          setShowList(true);
-          setVisualFocusIndex(firstIndex);
-          return;
-        case "Enter":
-        case "Escape":
-          event.preventDefault();
-          setValue("");
-          return;
-      }
-    };
-
-    const handleOpenedListKey = (event: KeyboardEvent) => {
-      const key = event.key;
-
-      switch (key) {
-        case "Up":
-        case "ArrowUp":
-          event.preventDefault();
-          const previousIndex =
-            visualFocusIndex === 0 || visualFocusIndex === null
-              ? options.length - 1
-              : visualFocusIndex - 1;
-          setVisualFocusIndex(previousIndex);
-          return;
-        case "Down":
-        case "ArrowDown":
-          event.preventDefault();
-          const nextIndex =
-            visualFocusIndex === options.length - 1 || visualFocusIndex === null
-              ? 0
-              : visualFocusIndex + 1;
-          setVisualFocusIndex(nextIndex);
-          return;
-        case "Enter":
-          event.preventDefault();
-          if (!(visualFocusIndex === null)) {
-            const currentValue = options[visualFocusIndex].name;
-            setValue(currentValue);
-          }
-          setShowList(false);
-          return;
-        case "Escape":
-          event.preventDefault();
-          if (showList) {
-            setShowList(false);
-          }
-          return;
-        default:
-          setVisualFocusIndex(null);
-      }
-    };
-
-    if (!showList) {
-      document.addEventListener("keydown", handleClosedListKey);
-
-      return () => {
-        document.removeEventListener("keydown", handleClosedListKey);
-      };
-    } else {
-      document.addEventListener("keydown", handleOpenedListKey);
-
-      return () => {
-        document.removeEventListener("keydown", handleOpenedListKey);
-      };
-    }
-  }, [isFocus, visualFocusIndex, showList]);
+  const { visualFocusIndex, setIsFocus } = useKeyboard({
+    options,
+    showList,
+    setShowList,
+    setValue,
+    setSelectedElementId,
+  });
 
   return (
     <div className="grid gap-2">
@@ -121,7 +40,10 @@ export const Combobox = ({
         onChange={(event) => setValue(event.target.value)}
         size={size}
         listName={listName}
-        onClick={() => setShowList(!showList)}
+        onClick={() => {
+          setIsFocus(true);
+          setShowList(!showList);
+        }}
         onFocus={() => setIsFocus(true)}
         onBlur={() => setIsFocus(false)}
         disabled={disabled}
